Create README.md
Minor description on possible reasons why the fast API is not connecting a path to correct that in your local environment.

An, a brief description on intended advancements later on with an explanation on how to achieve it. On your own Local environment does the example that I post Forward will be, of course. Tested, but it will not contain one of my actual cookies.