Add files via upload
Version 3.0.
changes made to the fast AP. I some pseudo code was replaced with actual logic
For the API handling specifically in regards to progress that is to say the progress of downloads and a face-lift to the index.html file.
Also included a optional variable. A cookies.txt file.